feat: Initial pairing of kafka pub-sub
Requires a kafka instance running in the background on the specified
details. Future instances will need to better document and abstract that
connection.

Introduces a close_or_die handler which should simplify the deferred
close process.

This initial version seems to hold all the logs until the end, which is
not the behaviour I'm looking for, along with being pretty slow. Future
versions will need to fix the dump-logs-at-end issue. The performance
may well be a high constant time delay introduced by the Kafka service,
but this can also be investigated in a later session.